refactor: separate fetch function
